LayoutTests/platform/mac/fast/forms/input-number-click-expected.txt


Test for the spin control behavior in a type=numnber input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS i.value is "4.14"
PASS i.value is "4.14"
PASS i.value is "0.14"
PASS i.value is "0.14"
PASS successfullyParsed is true

TEST COMPLETE
